requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# AWS Services
boto3>=1.29.7
botocore>=1.32.7

# Computer Vision & Object Detection
ultralytics>=8.1.0
pillow>=10.1.0

# Natural Language Processing
sumy>=0.11.0
nltk>=3.8.1

# Environment Variables
python-dotenv>=1.0.0

# Development Tools
pylint>=3.0.2

# Additional Dependencies (compatible with Python 3.12)
numpy>=1.26.2
opencv-python>=4.8.1.78
torch>=2.2.0
torchvision>=0.17.0


redis>=5.0.0
